Move the calendar coordinator into the coordinator module
Core requires DataUpdateCoordinator subclasses to live in `coordinator`
(pylint home-assistant-enforce-class-module), so the coordinator, its data
model and the event parsing move there, leaving the entity in `calendar`.

Also tick past the scan interval rather than exactly onto it in the
new-calendar test, which was racing the scheduled refresh.
